Move core classes to backend.py
